# Optimize kryo serialization