--- a/build.sbt
+++ b/build.sbt
@@ -34,7 +34,6 @@ val chillVersion = "0.9.2"
 val circeVersion = "0.9.1"
 val commonsIoVersion = "2.6"
 val commonsMath3Version = "3.6.1"
-val commonsPoolVersion = "2.5.0"
 val elasticsearch2Version = "2.1.0"
 val elasticsearch5Version = "5.5.0"
 val featranVersion = "0.1.27"
@@ -303,8 +302,7 @@ lazy val scioCore: Project = Project(
     "com.google.guava" % "guava" % guavaVersion,
     "com.google.protobuf" % "protobuf-java" % protobufVersion,
     "me.lyh" %% "protobuf-generic" % protobufGenericVersion,
-    "org.apache.xbean" % "xbean-asm5-shaded" % asmVersion,
-    "org.apache.commons" % "commons-pool2" %  commonsPoolVersion
+    "org.apache.xbean" % "xbean-asm5-shaded" % asmVersion
   )
 ).dependsOn(
   scioAvro,