### master
```
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] KryoAtomicCoderBenchmark.customDecode      avgt   80   235.947 ±  16.055  ns/op
[info] KryoAtomicCoderBenchmark.customEncode      avgt   80   804.571 ±  15.781  ns/op
[info] KryoAtomicCoderBenchmark.customKryoDecode  avgt   80   663.201 ±  17.165  ns/op
[info] KryoAtomicCoderBenchmark.customKryoEncode  avgt   80  1502.096 ±  19.122  ns/op
[info] KryoAtomicCoderBenchmark.javaDecode        avgt   80  7810.390 ± 110.794  ns/op
[info] KryoAtomicCoderBenchmark.javaEncode        avgt   80  5993.367 ± 108.613  ns/op
[info] KryoAtomicCoderBenchmark.kryoDecode        avgt   80   761.828 ±   8.562  ns/op
[info] KryoAtomicCoderBenchmark.kryoEncode        avgt   80  1603.633 ±  16.763  ns/op
```
### PR
``` 
[info] Benchmark                                  Mode  Cnt     Score     Error  Units
[info] KryoAtomicCoderBenchmark.customDecode      avgt   80   232.563 ±   5.763  ns/op
[info] KryoAtomicCoderBenchmark.customEncode      avgt   80   785.640 ±   5.570  ns/op
[info] KryoAtomicCoderBenchmark.customKryoDecode  avgt   80   192.968 ±   3.797  ns/op
[info] KryoAtomicCoderBenchmark.customKryoEncode  avgt   80  1085.145 ±  16.238  ns/op
[info] KryoAtomicCoderBenchmark.javaDecode        avgt   80  8033.934 ± 161.660  ns/op
[info] KryoAtomicCoderBenchmark.javaEncode        avgt   80  6498.303 ± 891.901  ns/op
[info] KryoAtomicCoderBenchmark.kryoDecode        avgt   80   313.482 ±   6.569  ns/op
[info] KryoAtomicCoderBenchmark.kryoEncode        avgt   80  1130.998 ±  16.154  ns/op
```

This was also test against a long running streaming pipeline to make sure we don't have memory leaks!